Add Dockerfile and Docker build workflow
This commit introduces a Dockerfile for building our application into a container. A GitHub Actions workflow has also been added for automatically building and pushing the Docker container when changes are pushed to master, or a pull request is opened against it. The workflow is set up for multi-platform builds.